Added a test for displaying the child clues.
This is the final pull request for iteration 5. IT contains code to parse the string for the parent clue list and save it to each corresponding clue. It then finds the child clues of each parent and saves it, which is displayed on the display clues page. Quite a few bugs were found during development, leaving this feature unfinshed.